backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
fastapi==0.116.1
filelock==3.18.0
firebase_admin==7.0.0
fonttools==4.59.0
fsspec==2025.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.168
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.35.0
